Avoid potential int overflow
Casting to type long before multiplying reduces the risk of arithmetic overflow.

+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/engine/ReactiveActionQueue.java
@@ -1120,7 +1120,7 @@ public class ReactiveActionQueue {
 
 			boolean sorted = false;
 
-			long maxIterations = latestBatches.size() * latestBatches.size();
+			long maxIterations = (long) latestBatches.size() * latestBatches.size();
 			long iterations = 0;
 
 			sort:
